+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Next
+
 -   feature: Allow user to implement custom toggle repository
 -   feat: Drop node 6 support (EOL)